refactor: Mutation for renewal check
Now using mutation from react-query for the renewal check which
happens every 6 hours. In line with earlier changes.
